refactor: extract connectionmanager setup
