+++ include/open62541/plugin/eventloop.h
@@ -10,7 +10,8 @@
 
 #include <open62541/types.h>
 #include <open62541/types_generated.h>
-#include <open62541/plugin/network.h>
+#include <open62541/util.h>
+#include <open62541/plugin/log.h>
 
 _UA_BEGIN_DECLS
